Studio: Add TitledBorders to Pixel Size dialogs.  Move PixelConfigExtraPanel to a better place.
